seperate auth route && add error handling in services
