+++ src/service/userService.ts
@@ -1,5 +1,5 @@
 import bcrypt from "bcryptjs";
-import jwt, { JwtPayload, Secret } from "jsonwebtoken";
+
 import { User } from "../interface/userInterfaces";
 import {
   getAllUsers,
@@ -10,10 +10,10 @@ import {
   deleteUser as deleteUserModel,
   generateNextUserId,
 } from "../model/userModel";
-import config from "../config";
 import ApiError from "../error/apiError";
-
-const { secretKey, refreshSecretKey } = config;
+import { StatusCodes } from "http-status-codes";
+import { BadRequestError } from "../error/badRequestError";
+import notFoundError from "../error/notFoundError";
 
 export const fetchUsers = (): User[] => getAllUsers();
 
@@ -27,13 +27,16 @@ export const createUser = async (
   email: string,
   password: string
 ): Promise<Omit<User, "password">> => {
+  if (fetchUserByEmail(email)) {
+    throw new BadRequestError("Email already in use");
+  }
   const hashedPassword = await bcrypt.hash(password, 10);
   const newUser: User = {
     id: generateNextUserId(),
     name,
     email,
     password: hashedPassword,
-    role: "",
+    role: "user",
   };
   const addedUser = addUser(newUser);
   const { password: _password, ...response } = addedUser;
@@ -46,7 +49,7 @@ export const updateUser = async (
 ): Promise<Omit<User, "password"> | null> => {
   const user = updateUserModel(id, updateData);
   if (!user) {
-    throw new ApiError(404, `User with ID ${id} not found`);
+    throw new notFoundError(`User with ID ${id} not found`);
   }
   const { password, ...response } = user;
   return response;
@@ -55,58 +58,7 @@ export const updateUser = async (
 export const deleteUser = (id: number): User | null => {
   const user = deleteUserModel(id);
   if (!user) {
-    throw new ApiError(404, `User with ID ${id} not found`);
+    throw new notFoundError(`User with ID ${id} not found`);
   }
   return user;
-};
-
-export const validateUserCredentials = async (
-  email: string,
-  password: string
-): Promise<User | null> => {
-  const user = getUserByEmail(email);
-  if (!user) return null;
-  const isValidPassword = await bcrypt.compare(password, user.password);
-  return isValidPassword ? user : null;
-};
-
-export const generateTokens = (user: User) => {
-  if (!secretKey || !refreshSecretKey) {
-    throw new Error("Secret keys are not defined");
-  }
-
-  const accessToken = jwt.sign(
-    { id: user.id, email: user.email },
-    secretKey as Secret,
-    { expiresIn: "1h" }
-  );
-  const refreshToken = jwt.sign(
-    { id: user.id, email: user.email },
-    refreshSecretKey as Secret,
-    { expiresIn: "7d" }
-  );
-  return { accessToken, refreshToken };
-};
-
-export const refreshAccessToken = (
-  refreshToken: string
-): { accessToken: string } | null => {
-  try {
-    const decoded = jwt.verify(
-      refreshToken,
-      refreshSecretKey as Secret
-    ) as JwtPayload;
-    const user = getUserByEmail(decoded.email);
-    if (!user) {
-      return null;
-    }
-    const accessToken = jwt.sign(
-      { id: user.id, email: user.email },
-      secretKey as Secret,
-      { expiresIn: "1h" }
-    );
-    return { accessToken };
-  } catch (error) {
-    return null;
-  }
 };
